requirements.txt
django
djangorestframework
django-filter
djangorestframework-bulk
gunicorn
django-heroku
python-dotenv
dj_database_url
djoser